+++ internal/git/filter.go
@@ -1,6 +1,9 @@
 package git
 
-import "regexp"
+import (
+	"log/slog"
+	"regexp"
+)
 
 func NewPathFilter(include, exclude, relaxed []*regexp.Regexp) PathFilter {
 	return PathFilter{
@@ -23,17 +26,19 @@ func (pf PathFilter) IsPathAllowed(path string) bool {
 
 	for _, pattern := range pf.exclude {
 		if pattern.MatchString(path) {
+			slog.Debug("File path is in the exclude list", slog.String("path", path), slog.Any("exclude", pf.exclude))
 			return false
 		}
 	}
 
 	for _, pattern := range pf.include {
 		if pattern.MatchString(path) {
+			slog.Debug("File path is in the include list", slog.String("path", path), slog.Any("include", pf.include))
 			return true
 		}
 	}
 
-	return false
+	return len(pf.include) == 0
 }
 
 func (pf PathFilter) IsRelaxed(path string) bool {